Fail Run() on missing language files
If no language files are available, there can be no
meaningful user interaction. Fail with returncode 5.

+++ run.go
@@ -46,6 +46,10 @@ func Run() int {
 	}
 	config.Variables["installerName"] = os.Args[0]
 	translator := NewTranslatorVar(config.Variables)
+	if translator == nil {
+		log.Println("No language files available")
+		return 5
+	}
 	installerTempPath := filepath.Join(os.TempDir(), "linux_installer")
 	defer os.RemoveAll(installerTempPath)
